+++ ValidCode/UsingGuReactive.cs
@@ -32,6 +32,21 @@ namespace ValidCode
                          .AsReadOnlyFilteredView(filter, leaveOpen: false);
         }
 
+        public static IReadOnlyView<IDisposable> M2(IObservable<IEnumerable<int>> source, Func<int, bool> filter)
+        {
+            if (source is null)
+            {
+                throw new ArgumentNullException(nameof(source));
+            }
+
+            if (filter is null)
+            {
+                throw new ArgumentNullException(nameof(filter));
+            }
+
+            return source.AsReadOnlyFilteredView(filter).AsMappingView(x => new Disposable(), onRemove:x => x.Dispose());
+        }
+
         public static void M2<T>(IObservable<IEnumerable<T>> source, Func<T, bool> filter)
         {
             using var view = source.AsReadOnlyView().AsReadOnlyFilteredView(filter, leaveOpen: false);